其他项目/nornir-vsr/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.30
pydantic==2.7.1
alembic==1.13.1
nornir==3.4.1
nornir-netmiko==1.0.1
nornir-utils==0.2.0
python-dotenv==1.0.1
setuptools==80.9.0
openpyxl==3.1.5
psycopg[binary]==3.1.12
asyncssh==2.14.2
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pyotp==2.9.0
APScheduler==3.10.4
